pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pillow>=12.1.0",
    "pydicom>=3.0.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
